Split up `animated_fox` example
